refactor(RacingCarService): 가독성 개선을 위해 메서드 이름 변경
- 테스트 포함
